add possibility for custom total amount

+++ app/src/main/java/com/example/potteryapp/database/Converters.kt
@@ -22,7 +22,7 @@ class Converters {
             if (str.isNullOrBlank())
                 break
             var items = str.split(',')
-            list.add(Item(items[0].toInt(), items[1], items[2].toInt(), items[3]))
+            list.add(Item(items[0].toInt(), items[1], items[2].toDouble(), items[3]))
         }
         return list
     }
