+++ app/src/main/java/com/example/potteryapp/views/FormulaDetailFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.example.potteryapp.databinding.FragmentFormulaDetailBinding
@@ -44,7 +45,8 @@ class FormulaDetailFragment : Fragment() {
         binding.button.setOnClickListener {
             val newTotal = binding.editTextTotalAmount.text.toString()
             val totalInt = Integer.parseInt(newTotal)
-            formula = vModel.createNewFormula(formula, totalInt)
+            formula = vModel.createNewFormula(formula, totalInt.toDouble())
+
             initList()
         }
     }
